% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far2.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Make the map by states}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of the state}

\item{year}{year for which injuries should be mapped}
}
\value{
This function returns the map of the given state
}
\description{
Visualize the data in the map
}
\examples{
\dontrun{
fars_map_state(43,2013)
fars_map_state(11,2014)
fars_map_state(3,2015)
}

}
