% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far2.R
\name{fars_read}
\alias{fars_read}
\title{File existancy}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a csv file that contains the data}
}
\value{
This function returns reading the dataset. If the dataset is not available it informs about the unavailability
}
\description{
Check if the file exists & let the user know that
}
\examples{
\dontrun{
fars_read(accident_2013)
fars_read(accident_2014)
fars_read(accident_2013, accident_2015)
}

}
