% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far2.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{value indicating the year/ years}
}
\value{
This function returns the "Month" & "Year". If incorrect year has entered it retunrs informing that the year is invalid.
}
\description{
Configure the years of the data set & check the validity
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(2013, 2014, 2015)
}

}
