% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far2.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of the years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{value indicating the year/ years}
}
\value{
This function returns the number of records grouping by year & month
}
\description{
Group the data by years & months together
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(2013, 2014)
}

}
