% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far2.R
\name{make_filename}
\alias{make_filename}
\title{Make the file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{value indicating the year}
}
\value{
This function returns making the file name
}
\description{
Making the file name in which including the year
}
\examples{
\dontrun{
make_filename(2013)
make_filename(2014)
}

}
